requirements.txt
asgiref==3.4.1
certifi==2023.7.22
charset-normalizer==3.2.0
Deprecated==1.2.13
Django==3.2.4
django-prometheus==2.1.0
django-redis==5.2.0
djangorestframework==3.13.1
idna==3.4
packaging==21.3
prometheus-client==0.12.0
psycopg2-binary==2.9.7
pyparsing==3.0.6
pytz==2021.3
redis==4.1.0
requests==2.31.0
sqlparse==0.4.2
urllib3==2.0.4
wrapt==1.13.3
